inferserver/qwen-vl-server/requirements.txt
transformers>=4.37
accelerate
qwen-vl-utils[decord]==0.0.8
autoawq
fastapi
uvicorn
modelscope
torch
Pillow
requests
